# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CSE341 API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "githubOAuth": {
      "type": "oauth2",
      "authorizationUrl": "https://github.com/login/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "user:email": "Access user email address"
      }
    }
  },
  "paths": {
    "/auth/github": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Initiate GitHub OAuth login",
        "responses": {
          "302": { "description": "Redirect to GitHub" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/auth/github/callback": {
      "get": {
        "tags": ["Authentication"],
        "summary": "GitHub OAuth callback",
        "responses": {
          "302": { "description": "Redirect after successful authentication" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/auth/logout": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Logout user",
        "responses": {
          "302": { "description": "Redirect to home" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/auth/status": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Check authentication status",
        "responses": {
          "200": {
            "schema": {
              "type": "object",
              "properties": {
                "authenticated": { "type": "boolean" },
                "user": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string" },
                    "displayName": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get all contacts",
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Contact" }
            }
          },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create a contact (Protected)",
        "security": [{ "githubOAuth": [] }],
        "parameters": [{
          "name": "body",
          "in": "body",
          "required": true,
          "schema": { "$ref": "#/definitions/ContactInput" }
        }],
        "responses": {
          "201": { "schema": { "$ref": "#/definitions/Contact" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized - authentication required" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact by ID",
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": { "schema": { "$ref": "#/definitions/Contact" } },
          "404": { "description": "Not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update contact (Protected)",
        "security": [{ "githubOAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/ContactInput" } }
        ],
        "responses": {
          "200": { "schema": { "$ref": "#/definitions/Contact" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized - authentication required" },
          "404": { "description": "Not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contact (Protected)",
        "security": [{ "githubOAuth": [] }],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Unauthorized - authentication required" },
          "404": { "description": "Not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get all products",
        "responses": {
          "200": {
            "schema": {
              "type": "array",
              "items": { "$ref": "#/definitions/Product" }
            }
          },
          "500": { "description": "Internal Server Error" }
        }
      },
      "post": {
        "tags": ["Products"],
        "summary": "Create a product (Protected)",
        "security": [{ "githubOAuth": [] }],
        "parameters": [{
          "name": "body",
          "in": "body",
          "required": true,
          "schema": { "$ref": "#/definitions/ProductInput" }
        }],
        "responses": {
          "201": { "schema": { "$ref": "#/definitions/Product" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized - authentication required" },
          "500": { "description": "Internal Server Error" }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Products"],
        "summary": "Get product by ID",
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "200": { "schema": { "$ref": "#/definitions/Product" } },
          "404": { "description": "Not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "put": {
        "tags": ["Products"],
        "summary": "Update product (Protected)",
        "security": [{ "githubOAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "type": "string" },
          { "name": "body", "in": "body", "required": true, "schema": { "$ref": "#/definitions/ProductInput" } }
        ],
        "responses": {
          "200": { "schema": { "$ref": "#/definitions/Product" } },
          "400": { "description": "Validation failed" },
          "401": { "description": "Unauthorized - authentication required" },
          "404": { "description": "Not found" },
          "500": { "description": "Internal Server Error" }
        }
      },
      "delete": {
        "tags": ["Products"],
        "summary": "Delete product (Protected)",
        "security": [{ "githubOAuth": [] }],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": true,
          "type": "string"
        }],
        "responses": {
          "204": { "description": "Deleted" },
          "401": { "description": "Unauthorized - authentication required" },
          "404": { "description": "Not found" },
          "500": { "description": "Internal Server Error" }
        }
      }
    }
  },
  "definitions": {
    "Contact": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "phone": { "type": "string" },
        "address": { "type": "string" },
        "dateOfBirth": { "type": "string", "format": "date" },
        "role": { "type": "string" }
      }
    },
    "ContactInput": {
      "type": "object",
      "required": ["firstName", "lastName", "email", "phone", "address", "dateOfBirth", "role"],
      "properties": {
        "firstName": { "type": "string" },
        "lastName": { "type": "string" },
        "email": { "type": "string" },
        "phone": { "type": "string" },
        "address": { "type": "string" },
        "dateOfBirth": { "type": "string", "format": "date" },
        "role": { "type": "string" }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "_id": { "type": "string" },
        "name": { "type": "string" },
        "description": { "type": "string" },
        "price": { "type": "number" },
        "category": { "type": "string" },
        "stock": { "type": "integer" },
        "sku": { "type": "string" },
        "brand": { "type": "string" }
      }
    },
    "ProductInput": {
      "type": "object",
      "required": ["name", "description", "price", "category", "stock", "sku", "brand"],
      "properties": {
        "name": { "type": "string" },
        "description": { "type": "string" },
        "price": { "type": "number" },
        "category": { "type": "string" },
        "stock": { "type": "integer" },
        "sku": { "type": "string" },
        "brand": { "type": "string" }
      }
    }
  }
}